Moving to utility class
